Add ResultDisplayTest
