--- src/test/java/seedu/address/ui/CommandBoxTest.java
@@ -1,152 +1,0 @@
-package seedu.address.ui;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-import java.util.ArrayList;
-
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Test;
-
-import guitests.guihandles.CommandBoxHandle;
-import javafx.scene.input.KeyCode;
-import seedu.address.logic.commands.CommandResult;
-import seedu.address.logic.commands.ListCommand;
-import seedu.address.logic.commands.exceptions.CommandException;
-
-public class CommandBoxTest extends GuiUnitTest {
-
-    private static final String COMMAND_THAT_SUCCEEDS = ListCommand.COMMAND_WORD;
-    private static final String COMMAND_THAT_FAILS = "invalid command";
-
-    private ArrayList<String> defaultStyleOfCommandBox;
-    private ArrayList<String> errorStyleOfCommandBox;
-
-    private CommandBoxHandle commandBoxHandle;
-
-    @BeforeAll
-    public void setUp() {
-        CommandBox commandBox = new CommandBox();
-        commandBoxHandle = new CommandBoxHandle(getChildNode(commandBox.getRoot(),
-                CommandBoxHandle.COMMAND_INPUT_FIELD_ID));
-        uiPartExtension.setUiPart(commandBox);
-
-        defaultStyleOfCommandBox = new ArrayList<>(commandBoxHandle.getStyleClass());
-
-        errorStyleOfCommandBox = new ArrayList<>(defaultStyleOfCommandBox);
-        errorStyleOfCommandBox.add(CommandBox.ERROR_STYLE_CLASS);
-    }
-
-    @Test
-    public void commandBox_startingWithSuccessfulCommand() {
-        assertBehaviorForSuccessfulCommand();
-        assertBehaviorForFailedCommand();
-    }
-
-    @Test
-    public void commandBox_startingWithFailedCommand() {
-        assertBehaviorForFailedCommand();
-        assertBehaviorForSuccessfulCommand();
-
-        // verify that style is changed correctly even after multiple consecutive failed commands
-        assertBehaviorForSuccessfulCommand();
-        assertBehaviorForFailedCommand();
-        assertBehaviorForFailedCommand();
-    }
-
-    @Test
-    public void commandBox_handleKeyPress() {
-        commandBoxHandle.run(COMMAND_THAT_FAILS);
-        assertEquals(errorStyleOfCommandBox, commandBoxHandle.getStyleClass());
-        guiRobot.push(KeyCode.ESCAPE);
-        assertEquals(errorStyleOfCommandBox, commandBoxHandle.getStyleClass());
-
-        guiRobot.push(KeyCode.A);
-        assertEquals(defaultStyleOfCommandBox, commandBoxHandle.getStyleClass());
-    }
-
-    @Test
-    public void handleKeyPress_startingWithUp() {
-        // empty history
-        assertInputHistory(KeyCode.UP, "");
-        assertInputHistory(KeyCode.DOWN, "");
-
-        // one command
-        commandBoxHandle.run(COMMAND_THAT_SUCCEEDS);
-        assertInputHistory(KeyCode.UP, COMMAND_THAT_SUCCEEDS);
-        assertInputHistory(KeyCode.DOWN, "");
-
-        // two commands (latest command is failure)
-        commandBoxHandle.run(COMMAND_THAT_FAILS);
-        assertInputHistory(KeyCode.UP, COMMAND_THAT_SUCCEEDS);
-        assertInputHistory(KeyCode.UP, COMMAND_THAT_SUCCEEDS);
-        assertInputHistory(KeyCode.DOWN, COMMAND_THAT_FAILS);
-        assertInputHistory(KeyCode.DOWN, "");
-        assertInputHistory(KeyCode.DOWN, "");
-        assertInputHistory(KeyCode.UP, COMMAND_THAT_FAILS);
-
-        // insert command in the middle of retrieving previous commands
-        guiRobot.push(KeyCode.UP);
-        String thirdCommand = "list";
-        commandBoxHandle.run(thirdCommand);
-        assertInputHistory(KeyCode.UP, thirdCommand);
-        assertInputHistory(KeyCode.UP, COMMAND_THAT_FAILS);
-        assertInputHistory(KeyCode.UP, COMMAND_THAT_SUCCEEDS);
-        assertInputHistory(KeyCode.DOWN, COMMAND_THAT_FAILS);
-        assertInputHistory(KeyCode.DOWN, thirdCommand);
-        assertInputHistory(KeyCode.DOWN, "");
-    }
-
-    @Test
-    public void handleKeyPress_startingWithDown() {
-        // empty history
-        assertInputHistory(KeyCode.DOWN, "");
-        assertInputHistory(KeyCode.UP, "");
-
-        // one command
-        commandBoxHandle.run(COMMAND_THAT_SUCCEEDS);
-        assertInputHistory(KeyCode.DOWN, "");
-        assertInputHistory(KeyCode.UP, COMMAND_THAT_SUCCEEDS);
-
-        // two commands
-        commandBoxHandle.run(COMMAND_THAT_FAILS);
-        assertInputHistory(KeyCode.DOWN, "");
-        assertInputHistory(KeyCode.UP, COMMAND_THAT_FAILS);
-
-        // insert command in the middle of retrieving previous commands
-        guiRobot.push(KeyCode.UP);
-        String thirdCommand = "list";
-        commandBoxHandle.run(thirdCommand);
-        assertInputHistory(KeyCode.DOWN, "");
-        assertInputHistory(KeyCode.UP, thirdCommand);
-    }
-
-    /**
-     * Runs a command that fails, then verifies that <br>
-     *      - the text remains <br>
-     *      - the command box's style is the same as {@code errorStyleOfCommandBox}.
-     */
-    private void assertBehaviorForFailedCommand() {
-        commandBoxHandle.run(COMMAND_THAT_FAILS);
-        assertEquals(COMMAND_THAT_FAILS, commandBoxHandle.getInput());
-        assertEquals(errorStyleOfCommandBox, commandBoxHandle.getStyleClass());
-    }
-
-    /**
-     * Runs a command that succeeds, then verifies that <br>
-     *      - the text is cleared <br>
-     *      - the command box's style is the same as {@code defaultStyleOfCommandBox}.
-     */
-    private void assertBehaviorForSuccessfulCommand() {
-        commandBoxHandle.run(COMMAND_THAT_SUCCEEDS);
-        assertEquals("", commandBoxHandle.getInput());
-        assertEquals(defaultStyleOfCommandBox, commandBoxHandle.getStyleClass());
-    }
-
-    /**
-     * Pushes {@code keycode} and checks that the input in the {@code commandBox} equals to {@code expectedCommand}.
-     */
-    private void assertInputHistory(KeyCode keycode, String expectedCommand) {
-        guiRobot.push(keycode);
-        assertEquals(expectedCommand, commandBoxHandle.getInput());
-    }
-}
